Track activity for scene playlist playback
