+++ app/src/main/java/com/github/damontecres/stashapp/playback/PlaylistScenesFragment.kt
@@ -43,7 +43,6 @@ class PlaylistScenesFragment :
                 moreOptionsButton,
                 listOf(debugToggleText, "Show playlist"),
             ) { position ->
-
                 if (position == 0) {
                     if (debugView.isVisible) {
                         debugView.visibility = View.GONE
